backend/requirements.txt
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
Flask-CORS==6.0.2
Flask-JWT-Extended==4.7.1
python-dotenv==1.2.1
requests==2.32.5